Squashed commit of the following:
commit b036c342cc9974f747cb83cb2f8ff886712e88e0
Merge: 8efea8754 ee174c1ac
Author: Maksym Malyhin <mmaksym@google.com>
Date:   Fri Jun 7 14:39:29 2019 -0400

    Merge branch 'mm/fis-master' into mm/fid-fid

commit 8efea8754fa44fb483969a7c8aabb0f141cebaf3
Author: Maksym Malyhin <mmaksym@google.com>
Date:   Fri Jun 7 14:24:43 2019 -0400

    Run ./scripts/style.sh

commit f44f475fd29b93dbbcc44b9079687f7c94bf33d4
Author: Maksym Malyhin <mmaksym@google.com>
Date:   Fri Jun 7 14:21:21 2019 -0400

    FIRInstallationsTests update to mock installationsIDController

commit 0d3b761ad8359367a70f8f53d9a6199b838fe7dc
Author: Maksym Malyhin <mmaksym@google.com>
Date:   Fri Jun 7 11:59:13 2019 -0400

    FID 22 chars long.

commit 84feb3ef5d8faf2903605e5e49ba3ad700ec4723
Author: Maksym Malyhin <mmaksym@google.com>
Date:   Fri Jun 7 11:40:08 2019 -0400

    FirebaseIDController tests

commit 5b4d866728b879b132b6248b36cc8b99c24c99ce
Author: Maksym Malyhin <mmaksym@google.com>
Date:   Fri Jun 7 10:47:21 2019 -0400

    FIRInstallationsIDController draft

commit 52b5cafe1b2a606cf6c9d5731450e40977a7e324
Author: Maksym Malyhin <mmaksym@google.com>
Date:   Thu Jun 6 16:19:16 2019 -0400

    [WIP] FID

commit aad6d17cb2ede689c884c582c2e9d12c18cb9ca0
Author: Maksym Malyhin <mmaksym@google.com>
Date:   Thu Jun 6 15:47:22 2019 -0400

    23 chars long FID

commit 9bb7ed3caa846a5717505a3157c62936b0bbaa23
Author: Maksym Malyhin <mmaksym@google.com>
Date:   Thu Jun 6 15:44:35 2019 -0400

    Generate 22 chars FID
