+++ FirebaseInstallations.podspec
@@ -38,7 +38,7 @@ Pod::Spec.new do |s|
     s.dependency 'GoogleUtilities/UserDefaults', '~> 6.2'
 
     s.test_spec 'unit' do |unit_tests|
-      unit_tests.source_files = base_dir + 'Tests/*.[mh]'
+      unit_tests.source_files = base_dir + 'Tests/**/*.[mh]'
       unit_tests.requires_app_host = true
       unit_tests.dependency 'OCMock'
     end
