+++ FirebaseInstallations/Source/Library/FIRInstallationsItem.h
@@ -17,7 +17,7 @@
 // TODO: Add short docs to the undocumented API.
 #import <Foundation/Foundation.h>
 
-typedef NS_ENUM(NSInteger, FIRInstallationsStatus);
+#import "FIRInstallationsStatus.h"
 
 @class FIRInstallationsStoredItem;
 @class FIRInstallationsStoredAuthToken;
@@ -48,6 +48,8 @@ NS_ASSUME_NONNULL_BEGIN
 - (NSString *)identifier;
 
 + (NSString *)identifierWithAppID:(NSString *)appID appName:(NSString *)appName;
+
++ (NSString *)generateFID;
 
 @end
