+++ FirebaseInstallations/Source/Library/InstallationsStore/FIRInstallationsStoredItem.h
@@ -17,29 +17,11 @@
 // TODO: Add short docs to the undocumented API.
 #import <Foundation/Foundation.h>
 
+#import "FIRInstallationsStatus.h"
+
 @class FIRInstallationsStoredAuthToken;
 
 NS_ASSUME_NONNULL_BEGIN
-
-/**
- * The enum represent possible states of the installation ID.
- *
- * WARNING: The enum is stored to Keychain as a part of `FIRInstallationsStoredItem`. Modification
- * of it can lead to incompatibility with previous version. Any modification must be evaluated and,
- * if it is really needed, the `storageVersion` must be bumped and proper migration code added.
- */
-typedef NS_ENUM(NSInteger, FIRInstallationsStatus) {
-  /** Represents either an initial status when a FIRInstallationsItem instance was created but not
-   * stored to Keychain or an undefined status (e.g. when the status failed to deserialize).
-   */
-  FIRInstallationStatusUnknown,
-  /// The Firebase Installation has not yet been registered with FIS.
-  FIRInstallationStatusUnregistered,
-  /// #CreateInstallation request to FIS server-API is in progress.
-  FIRInstallationStatusRegistrationInProgress,
-  /// The Firebase Installation has successfully been registered with FIS.
-  FIRInstallationStatusRegistered,
-};
 
 /**
  * The class is supposed to be used by `FIRInstallationsStore` only. It is required to
